Made file buffer consistent over files to avoid reallocation. Increased buffer size

+++ searchquery.h
@@ -5,7 +5,7 @@
 
 class SearchQuery {
     private:
-        short* bcTable[ALPHABET_SIZE];
+        int* bcTable[ALPHABET_SIZE];
 
         void createBCTable();
